+++ add-on/utils.py
@@ -30,3 +30,13 @@ def format_prompt_list(prompt_list, placeholder_dict, language_list=None):
             prompt = prompt.replace(f"#language#", language_list[index])
         formatted_prompt_list.append(prompt)
     return formatted_prompt_list
+
+
+def prompt_html(prompt, color):
+    return f"<br><br><font color='{color}'>Prompt:<br>{prompt}</font><br><br>"
+
+
+def field_value_html(field_value_list, color):
+    field_value_str = '<br>'.join(field_value_list)
+
+    return f"<font color='{color}'>Choosen values:</font><br>{field_value_str}"